pyproject.toml
[project]
name = "creator-pro-analysis"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "httpx>=0.27",
  "pydantic>=2.9",
  "python-dotenv>=1.0",
  "supabase>=2.29.0",
  "anthropic>=0.97.0",
  "openai>=2.32.0",
  "youtube-transcript-api>=1.2.4",
  "google-api-python-client>=2.194.0",
  "yt-dlp>=2026.3.17",
]
